# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealthDesk Plus API",
        "description": "Comprehensive healthcare management system for doctors, patients, appointments, prescriptions, and medical records.",
        "contact": {
            "name": "HealthDesk Plus Support",
            "email": "support@healthdeskplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.healthdeskplus.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/ai/profile/analyze": {
            "get": {
                "tags": [
                    "Profile Completeness"
                ],
                "summary": "Analyze profile completeness",
                "description": "Get AI-powered analysis of profile completeness with suggestions",
                "operationId": "f02572a9c96ddbed1ba1b44806709caf",
                "responses": {
                    "200": {
                        "description": "Profile analysis completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "completion_percentage": {
                                                    "type": "integer",
                                                    "example": 75
                                                },
                                                "basic_analysis": {
                                                    "properties": {
                                                        "completion_percentage": {
                                                            "type": "integer",
                                                            "example": 75
                                                        },
                                                        "total_fields": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "completed_fields": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "missing_fields": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "critical_missing": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "important_missing": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "ai_suggestions": {
                                                    "properties": {
                                                        "priority_actions": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "engagement_suggestions": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "professional_development": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "impact_analysis": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "priority_actions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "action": {
                                                                "type": "string",
                                                                "example": "Complete your full name"
                                                            },
                                                            "field": {
                                                                "type": "string",
                                                                "example": "name"
                                                            },
                                                            "priority": {
                                                                "type": "string",
                                                                "example": "high"
                                                            },
                                                            "estimated_time": {
                                                                "type": "string",
                                                                "example": "2-5 minutes"
                                                            },
                                                            "impact": {
                                                                "type": "string",
                                                                "example": "Critical for profile completion"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "estimated_time_to_complete": {
                                                    "type": "string",
                                                    "example": "30 minutes"
                                                },
                                                "impact_score": {
                                                    "type": "integer",
                                                    "example": 85
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}